ai_service/requirements.txt
torch
numpy
rasterio
flask
python-dotenv
requests
matplotlib

